requirements.txt
google-generativeai==0.3.2
sentence-transformers==2.2.2
chromadb==0.4.15
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
redis==5.0.1
